server_py/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
motor==3.7.1
pymongo==4.12.1
google-generativeai==0.8.5
pydantic==2.11.3
python-dotenv==1.1.0
slowapi==0.1.9
